Separated form_tests with form_data_tests and added a few more
