+++ apps/site/api/tests/__init__.py
@@ -3,6 +3,7 @@ from localground.apps.site.api.tests.admin_tests import *
 from localground.apps.site.api.tests.association_tests import *
 from localground.apps.site.api.tests.audio_tests import *
 from localground.apps.site.api.tests.form_tests import *
+from localground.apps.site.api.tests.form_data_tests import *
 from localground.apps.site.api.tests.field_tests import *
 from localground.apps.site.api.tests.homepage_tests import *
 # from localground.apps.site.api.tests.layer_tests import \
